# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API - Carteira Financeira",
        "description": "Documentação da API de autenticação, cadastro e operações financeiras.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:80",
            "description": "Servidor de desenvolvimento"
        }
    ],
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Cadastro"
                ],
                "summary": "Registra um novo usuário",
                "operationId": "ed7703a18fb888779fc84a24730577f8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "cpf_cnpj",
                                    "password",
                                    "password_confirmation",
                                    "street",
                                    "number",
                                    "neighborhood",
                                    "city",
                                    "state",
                                    "zipcode"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João da Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "joao@email.com"
                                    },
                                    "cpf_cnpj": {
                                        "type": "string",
                                        "example": "00000000000"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Senha123!"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "Senha123!"
                                    },
                                    "street": {
                                        "type": "string",
                                        "example": "Rua A"
                                    },
                                    "number": {
                                        "type": "string",
                                        "example": "123"
                                    },
                                    "neighborhood": {
                                        "type": "string",
                                        "example": "Centro"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "São Paulo"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "SP"
                                    },
                                    "zipcode": {
                                        "type": "string",
                                        "example": "12345-678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário registrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "João da Silva"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "Bearer eyJ.."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "500": {
                        "description": "Erro interno no servidor"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realiza o login de um usuário",
                "operationId": "8200ac786f4609dded9b146bad6d8d48",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "joao@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Senha123!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login bem-sucedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "João da Silva"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "Bearer eyJ..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Faz logout do usuário autenticado",
                "operationId": "ddcc690a7d51ed62a6dfe3158c12ed00",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logout realizado com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao realizar logout"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transactions/transfer": {
            "post": {
                "tags": [
                    "Transações"
                ],
                "summary": "Realiza uma transferência entre usuários",
                "operationId": "5e29baa45ec55960af3d0d33b75e5b21",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "payee_id"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    },
                                    "payee_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "metadata": {
                                        "type": "object",
                                        "example": {
                                            "descricao": "pagamento de serviço"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transferência realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transferência realizada com sucesso"
                                        },
                                        "transaction": {
                                            "type": "object"
                                        },
                                        "balance": {
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Saldo insuficiente ou bloqueado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transactions/deposit": {
            "post": {
                "tags": [
                    "Transações"
                ],
                "summary": "Realiza um depósito na carteira do usuário autenticado",
                "description": "Usuários autenticados podem realizar depósitos. O saldo será atualizado.",
                "operationId": "d8a54cd1da86a20423c0c5a13640df7a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Depósito realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Depósito realizado com sucesso."
                                        },
                                        "transaction": {
                                            "type": "object"
                                        },
                                        "balance": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 1200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation errors"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro interno ao processar depósito."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insira o token JWT no formato: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Autenticação",
            "description": "Endpoints de registro e login"
        },
        {
            "name": "Transações",
            "description": "Operações de saldo e transferências"
        },
        {
            "name": "Cadastro",
            "description": "Cadastro"
        }
    ]
}